[modify] cpplint检查格式化

+++ src/base/thread/utils.cc
@@ -1,6 +1,7 @@
 #include "src/base/thread/utils.h"
 
 #include <sstream>
+#include <string>
 #include <thread>
 
 namespace base {
@@ -8,7 +9,7 @@ uint64_t GetCurrentThreadId() {
   std::ostringstream stream;
   stream << std::this_thread::get_id();
   std::string stid = stream.str();
-  unsigned long long tid = std::stoull(stid);
+  uint64_t tid = std::stoull(stid);
   return tid;
 }
 }  // namespace base
